refactor: tidy the account-security pass without changing behavior
- Tests move next to their subjects. `secondary_hardening_test.go` was a bin whose
  name only made sense relative to the PR that created it; the webhook, origin,
  limiter, and key-scope cases now sit beside the code they describe, and the
  mint-scope assertion folds into the existing TestMintTailoringKey (which already
  had a minter fake, so the second one was redundant).
- ChangePassword read the caller's hash by id → email → hash. One query
  (PasswordHash) says what it means and drops a round-trip.
- The Telegram enable condition was two ifs sharing a duplicated prefix, differing
  only by == vs != on the secret; one gate with an if/else reads as the decision it is.
- AuthDialog's three near-identical mode switches collapse into switchMode().
